revert: "feat: cloudwatch logs transformer support" (#34984)
Reverts aws/aws-cdk#34863

This PR is being reverted because it introduced new enums containing quotes, which exposed a bug in our construct-metadata-updater package. The bug caused the release build to fail.

Since we need to ship a P0 fix in the upcoming release, we're temporarily reverting this PR while we develop a proper long-term solution for handling quoted enum values in the construct-metadata-updater.