fix(autoscaling): add an option to run validation when migrating to launch template (#34832)
This patch add an option to run validation when migrating to launch template, which checks if UpdatePolicy is specified.

### Issue # (if applicable)

Closes #34283.

### Reason for this change

To make migration to Launch Template safer.

### Description of changes

This change introduces an option flag to check whether or not UpdatePolicy is specified. As enforcing UpdatePolicy, existing EC2 instances based on Launch Configuration are terminated, and new EC2 instances based on Launch Template are launched, which always references IAM instance profile defined by Launch Template.

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

Added an unit test and integration test.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*