fix(app-staging-synthesizer): custom bootstrap qualifier is not passed through to deployment role name (#35118)
### Issue #28195 

Closes #28195 

### Reason for this change

When passing `bootstrapQualifier` in props to `AppStagingSynthesizer.defaultResources`, the synthesizer would still use the default qualifier `'hnb659fds'` when looking for bootstrap roles.

### Description of changes


`BootstraplessSynthesizer` is modified to take `qualifier` as an optional argument (if not provided, default bootstrap qualifier 'hnb659fds' is used).

The `bootstrapqualifier` is passed to `BootstraplessSynthesizer`, which is called in `AppStagingSynthesizer.defaultResources()`.

These changes ensure that calls to `AppStagingSynthesizer.defaultResources` using the `bootstrapQualifier` will use the qualifier in the deployment and CloudFormation execution roles instead of the default qualifier 'hnb659fds'.

### Describe any new or updated permissions being added

None.


### Description of how you validated changes

Added unit tests for:
- `BootstraplessSynthesizer`, which now optionally takes `qualifier` as an option
- `AppStagingSynthesizer`, which passes `qualifier` to `BootstraplessSynthesizer`

Tested by hand in a personal dev account.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*