fix(codebuild): support Windows Server Core 2022 image with on-demand capacity (#35152)
### Issue # (if applicable)

N/A

### Reason for this change

When support for the Windows Server Core 2022 image in AWS CodeBuild was initially added in #29754, the image type was not yet available for use with on-demand capacity.

This has now changed apparently (and tested), so we can remove the validation preventing use.
Additional removed a future looking test, since it is bad practices to add validations for the future without clearly documented evidence that this is going to happen.

I've searched the current docs, and there's no note whatsoever that Windows Server Core 2022 images are not supported with on-demand capacity.

### Description of changes

- Added Windows Server Core 2022 image support to the CodeBuild project configuration
- Updated integration tests to include the new Windows Core 2022 image variant
- Updated test snapshots to reflect the new image option

### Describe any new or updated permissions being added

No new or updated IAM permissions are required for this change.

### Description of how you validated changes

- Updated integration tests to include Windows Server Core 2022 image
- Executed updated integration test to validate the configuration and updated test snapshots 
- Existing unit tests continue to pass

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*