feat(ec2): support the new `ServiceRegion` property for `AWS::EC2::VPCEndpoint` (#35025)
### Issue # (if applicable)

Closes #32785 

### Reason for this change

Supporting the new L1 property in the L2 construct. `AWS::EC2::VPCEndpointService` support was added in #33959, so this should complete support for cross-region vpc private links.

### Description of changes

Added a new L2 prop - `serviceRegion` - which is of type `string`. It gets passed to the L1 `ServiceRegion` property when present.

### Describe any new or updated permissions being added

None

### Description of how you validated changes

Added unit tests

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*