chore(flags): feature flag report contains feature info for all library versions
This change goes together with
https://github.com/aws/aws-cdk-cli/pull/782, to undo some decisions made
in https://github.com/aws/aws-cdk/pull/35108/files and
https://github.com/aws/aws-cdk-cli/pull/742.

We used to store `unconfiguredBehavesLike` values for both `v1` and `v2`
(and `v3`, `v4`, etc) into the feature flag report, but we only needed to
store the value for the *current* library version. We are now defining
the "v2" field to hold the current version information, and are updating
the feature flag generation code to not emit `v1` data, and stay correct
if we ever start releasing `v3`.