chore(flags): feature flag report contains feature info for all library versions (#35215)
This change goes together with
https://github.com/aws/aws-cdk-cli/pull/782, to undo some decisions made in https://github.com/aws/aws-cdk/pull/35108/files and https://github.com/aws/aws-cdk-cli/pull/742.

We used to store `unconfiguredBehavesLike` values for both `v1` and `v2` (and `v3`, `v4`, etc) into the feature flag report, but we only needed to store the value for the *current* library version. We are now defining the "v2" field to hold the current version information, and are updating the feature flag generation code to not emit `v1` data, and stay correct if we ever start releasing `v3`.

This is labeled as a `chore` and not a `fix` because it doesn't really fix user-visible behavior.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*