fix(s3-deployment): list tokens in Source.jsonData are not resolved. (#35169)
### Issue #35145 

Closes #35145.

### Reason for this change

The recent fix to Source.jsonData did not work for list tokens such as SSM string list parameters. There was a mistake with the fix that assumed `stack.toJsonString` handled those, when in fact the tokens were serialized by `JSON.stringify`, breaking the token itself. `Token.isUnresolved` does not check all object properties for tokens, so it considered the object resolved and passed it to the regular JSON serializer.

### Description of changes

I've rewritten the JSON and YAML serializer such that it will use `stack.toJsonString` whenever it encounteres tokens in any of the object's children. One exception this is if the key in an object is a token, the token is assumed to be a string (because it has to in JSON) and is passed directly.

### Describe any new or updated permissions being added

N/A.

### Description of how you validated changes

- Unit and Integration tests added.
- Via manual inspection of stack in AWS Console, I checked that the JSON result was valid.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*