fix(signer): not correctly passing signingProfileName to CfnSigningProfile  (#35033)
### Issue # (if applicable)
Closes #35030.

### Reason for this change
The `signingProfileName` property set in the AWS CDK's Signer module L2 construct is not being properly propagated to the generated L1 CloudFormation template. This causes AWS Signer profiles to be created without the specified name, which leads to unexpected behavior and difficulty in managing resources.

### Description of changes
- Modified the `SigningProfile` class constructor to pass `this.physicalName` to the L1 `CfnSigningProfile` resource's `profileName` property
- Updated unit tests to explicitly verify that the `ProfileName` property is correctly included in the CloudFormation template when `signingProfileName` is provided, and absent when not provided

These changes ensure that when a user specifies a `signingProfileName`, it will be included in the generated CloudFormation template as the `ProfileName` property. The fix maintains backward compatibility and doesn't introduce any breaking changes to the API.

### Describe any new or updated permissions being added
No new or updated IAM permissions are needed for this change.

### Description of how you validated changes
- Updated existing unit tests to verify the fix
- Verified that the existing integration test for SigningProfile with a name works correctly
- Manually tested with a simple test application that uses SigningProfile with signingProfileName

### Why we need a FF to avoid BC

If we had specified a `signingProfileName` in their SigningProfile construct, the CloudFormation template will now include this name as the ProfileName property.

Previously, this name was ignored in the CloudFormation template, causing CloudFormation to auto-generate a name for the signing profile.

Upon deployment with the updated CDK version, CloudFormation will detect this as a change and will:
Create a new signing profile with the specified name.

Delete the old signing profile that had an auto-generated name

This replacement of the signing profile can break existing references to the profile and disrupt workflows that depend on the existing profile hence this PR introduced a FF to avoid this.


### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----
*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*