chore(lambda-nodejs): bump pnpm version in Dockerfile for AWS Lambda Node.js (#34537)
### Issue # (if applicable)

Closes #28318

### Reason for this change

The previous version of pnpm (7.30.5) used in the Dockerfile is no longer supported.
Therefore, I upgraded pnpm to the latest version 8.x.
Additionally, this update addresses the bug reported in issue #28318.

### Description of changes

Updated the version of pnpm installed in the Dockerfile from 7.30.5 to 8.15.9, as the previous version is no longer supported.


### Description of how you validated changes

I ran the unit tests for aws-lambda-nodejs.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*