chore(glue-alpha): remove documentation referring to .of() access to the constructors of WorkerType and JobType (#34734)
### Issue # (if applicable)

Closes part of #33356.

### Reason for this change

Recent changes to the aws-glue-alpha package removed WorkerType.of() and JobType.of() but the documentation remained  leading to some confusion when investigating compilation errors. After reviewing the PR that initially removed the methods I believe this is the preferred direction for the package. 

### Description of changes

Removed the comments referring to these methods.

If it would be preferable to re add the WorkerType.of() and JobType.of() methods I'm happy to adjust this PR accordingly.

### Describe any new or updated permissions being added

N/A


### Description of how you validated changes

Removed comments, ensured that yarn build still completed

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*