docs(cloudwatch): correct default value for YAxisProps, refactor GaugeWidget code (#35333)
### Issue #35262

Closes #35262.

### Reason for this change



`YAxisProps` indicated a default value of 0 for `min`, and No maximum value for `max`. This does not hold true for any class that uses this property.

If the default value of 0 for min were true, then omitting it should produce the same behavior in graphs as if `min: 0` were assigned. This is not the case. See https://github.com/cdklabs/construct-hub/pull/1775 for a counterexample. 

#### [AlarmWidget](https://github.com/aws/aws-cdk/blob/ec18b8300e94330587b36cfe7ce3f05b63b42d0f/packages/aws-cdk-lib/aws-cloudwatch/lib/graph.ts#L102)
The default is [no minimum or maximum for the y-axis](https://github.com/aws/aws-cdk/blob/ec18b8300e94330587b36cfe7ce3f05b63b42d0f/packages/aws-cdk-lib/aws-cloudwatch/lib/graph.ts#L130-L131), as we do NOT pass min and max by default.

#### [GaugeWidget](https://github.com/aws/aws-cdk/blob/ec18b8300e94330587b36cfe7ce3f05b63b42d0f/packages/aws-cdk-lib/aws-cloudwatch/lib/graph.ts#L180)
The default is [0 as the `min` and 100 as the `max`](https://github.com/aws/aws-cdk/blob/ec18b8300e94330587b36cfe7ce3f05b63b42d0f/packages/aws-cdk-lib/aws-cloudwatch/lib/graph.ts#L278-L279).

#### [GraphWidget](https://github.com/aws/aws-cdk/blob/ec18b8300e94330587b36cfe7ce3f05b63b42d0f/packages/aws-cdk-lib/aws-cloudwatch/lib/graph.ts#L312)
Here, we do NOT pass `min` and `max` by default, and they are automatically configured in CloudWatch if not provided.

### Description of changes



Corrected the `@default` to "Auto" for both `min` and `max`. Simplified the logic for passing `YAxisProps` in GaugeWidget, as there was previously dead code.

### Describe any new or updated permissions being added


None.

### Description of how you validated changes


Ran unit tests, which have sufficient coverage for the change to GaugeWidget.

No testing needed for documentation change.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*