chore(ci): get pr number and sha from github context (#35405)
### Issue # (if applicable)

Related to #35268 as the previous fix didn't fully work


### Reason for this change

The Pr linter workflow does not reach the code of the linter itself because it stop on error when being triggered at the end of a build, see the [following run](https://github.com/aws/aws-cdk/actions/runs/17435099055/job/49502649374) for example.

### Description of changes
Continue if `'Download workflow_run artifact'` to get the PR number and the pr sha from the Github context.


### Description of how you validated changes

Ran on my fork of the repo (canceling the build workflow retriggers this one), see the following run for example: https://github.com/leonmk-aws/aws-cdk/actions/runs/17435786087/job/49505111010

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*