chore: change PYTHON_LATEST from static versioning to a function (#35461)
### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change
Having a static readonly variable to manage the latest runtime version has proven to be challenging on NodeJS side, and a dynamic approach is preferred. 

### Description of changes
Changing the recently introduced `PYTHON_LATEST` static runtime to be a function. For now it will still deliver the latest python version available, but it will be changed into a dynamic approach later, similar to `determineLatestNodeRuntime`

### Describe any new or updated permissions being added

NA


### Description of how you validated changes

Updated existing unit tests

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*