fix(aws-ecs): allow empty placementStrategies on EC2Service
Same fix as #30382 but for `placementStrategies`. Apparently this was
not done at that time because CloudFormation [previously did not have
the appropriate
semantics](https://github.com/aws/aws-cdk/pull/27572#issuecomment-1766287866)
for setting PlacementStrategies to the empty array as it did for
PlacementConstraints, but that is no longer the case. The CloudFormation
docs
[state](https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-ecs-service.html#cfn-ecs-service-placementstrategies):
> To remove this property from your service resource, specify an empty PlacementStrategy array.

And indeed that works, with this CDK change applied.