fix(kms): cannot access aliasTargetKey on an Alias imported by Alias.fromAliasName (#35521)
If an Alias is ~~imported~~ referenced using `fromAliasName`, it doesn't have an `IKey` object it's associated with.

If an Alias is then used in place of a Key and the consuming site uses the `IKey.keyRef` interface to obtain the key ARN, we should return the Alias ARN and not try to delegate to the underlying key... because there is none!

Closes #35520 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*