fix(cloudformation-include): aws::novalue type validation error for non-string properties (#35425)
### Issue # (if applicable)
 
Closes #18420 
 
### Reason for this change
 
When we try to include a Cloudformation template that uses `{ Ref: 'AWS::NoValue' }` for a boolean-typed property, we get an error like:
 
`Error: Expected 'true' or 'false' for boolean value, got: '${Token[AWS.NoValue.7]}'`
 
This problem occurs because the CFN parser uses the `Aws.NO_VALUE` constant (which is defined as a string). CDK sees a string value for a boolean property and attempts to convert the string to boolean values. This results in the error we see above. 
 
### Description of changes
 
This change modifies the CFN parser in cloudformation-include to return a token value instead of a direct string constant for AWS::NoValue references. This makes it an IResolvable token that works with any property type. The parser then, in turn, ignores the default error being thrown whenever a non-string is passed. 
 
### Describe any new or updated permissions being added
 
No new IAM permissions required
 
### Description of how you validated changes
 
A unit test is added to check if an IResolvable token is returned for AWS::NoValue. An integration test deploys a Cloudformation stack with `AWS::NoValue` in a boolean property. Both tests pass. An assertion isn't needed here since the deployment of the infrastructure indicates that the test succeeded. 
 
### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)
 
----
 
*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*