fix(kms): Alias reference incorrectly resolves to underlying Key
Because `IAlias` extends `IKey`, `IAlias` inherits the new `keyRef:
KeyReference`.

It was originally implemented by referencing the underlying key, but
that's wrong: it should reference the alias itself, but using the field
names it shares with `IKey`.

In fact, we should have introduced a new interface like `IKeyLike` to do
this job, but instead we overloaded `IKey` to behave like a hypothetical
`IKeyLike`, and `IKeyRef` now inherits this double duty.

Therefore, we now make the `IKeyRef` implementation of `Alias` behave
like a `IKeyLikeRef`, and satisfy the contract using its own fields.