fix(cloudfront): `Function` ARN reference changed from GetAtt to Ref
When copying the underlying `CfnFunction`'s implementation of
`IFunctionRef`, the source of the `functionArn` value changed from a `{
GetAtt }` to a `{ Ref }`.

Both expression return the same value, but if the ARN is exported to
another stack, CloudFormation will reject any change to the expression
under the assumption that `different expression ⇒ different value`, and
that's not allowed.

Override the implementation for the L2, returning the same expression
as previously.