fix(rds): cannot use connections in rds.DatabaseInstance.fromLookup (#35193)
### Issue # (if applicable)

Closes #35192 

### Reason for this change
This change fixes an issue in DatabaseInstance.fromLookup() where port values aren't properly converted from strings to numbers. 

This causes validation errors when setting up access permissions using the connections interface, as  CfnSecurityGroupIngressProps requires port values to be numbers, not strings.


### Description of changes
Modified `instance.ts` in the `fromLookup` method to convert the port string to a number.



### Describe any new or updated permissions being added
N/A



### Description of how you validated changes
Add a unit test.


### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*