ci: add codebuild deployment integ runner PR workflow. (#35459)
### Reason for this change

- Deploying integration tests has been time-consuming and expensive for both maintainers and contributors of aws-cdk. Snapshot update and deployment may take from 10 minutes to hours depending on the snapshots updates made for each pull request. This can make any changes to aws-cdk, no matter how simple they are, be needlessly time-consuming and error-prone.
- AWS CDK's PR workflows does not validate that the snapshots of contributors are _deployable_. Although this is not common, snapshots updates can be made via a dry run, and there is a possibility that those snapshot updates won't be deployable. No PR Github action at the moment prevents this scenario.

### Description of changes

This adds a PR action which will use our Internal environment to test snapshot updates by attempting to deploy them. This uses the same service used by aws-cdk-cli (called Atmosphere) to deploy only the snapshot updates for the PR. This PR will need manual approval by a maintainer to run for the PR to work.

Mergify rules will also be updated such that if a request to run the integ tests are made, then this workflow should pass successfully in order for the PR to be merged.

### Description of how you validated changes

This workflow was validated with my fork of aws-cdk. See this example run: https://github.com/Abogical/aws-cdk/actions/runs/17458740897/job/49578262885

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*