feat(logs): add parameter to allow metric filter on transformed logs (#35359)
### Issue # (if applicable)

Not related to any issue.

### Reason for this change


It was not possible to pass the argument 'applyOnTransformedLogs' to the class MetricFilter, in order to create MetricFilter, Filter and Alarm. 

### Description of changes

Since MetricFilter makes use of CfnMetricFilter, which allows the parameter 'applyOnTransformedLogs', this change adds the parameter 'applyOnTransformedLogs' to MetricFilterOptions, which is directly passed to the CfnMetric class. 

### Describe any new or updated permissions being added

No permission is added or updated.

### Description of how you validated changes

I tested by hand. 

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*